fix chroma routes

+++ services/chroma.py
@@ -1,27 +1,39 @@
-from typing import Dict, List, Optional
+from typing import Dict, List, Optional, Any
 from datetime import datetime
 
 import chromadb
+from chromadb.errors import NotFoundError
+from chromadb.utils import embedding_functions
+
+sentence_transformer_ef = embedding_functions
 
 
 class ChromaService:
-    def __init__(self, embedding_model: str) -> None:
+    def __init__(self, embedding_model: str = "all-mpnet-base-v2") -> None:
         self.client = chromadb.PersistentClient()
         self.embedding_model = embedding_model
 
-    def create_collection(self, name, metadata):
-        chroma_collection = self.client.get_or_create_collection(name, metadata=metadata)
-        return chroma_collection
-
-    def get_collection(self, name: str):
-        return self.client.get_collection(name)
+    def get_or_create(self, name: str, metadata: Optional[Dict[str, str]] = None):
+        return self.client.get_or_create_collection(name=name, metadata=metadata, embedding_function=sentence_transformer_ef)
 
     def delete_collection(self, name: str):
-        self.client.delete_collection(name)
+        try:
+            collection = self.client.get_collection(name)
+            if collection:
+                self.client.delete_collection(name)
+        except NotFoundError:
+            raise NotFoundError(f"Collection '{name}' not found")
 
     def add_document_to_collection(self, name: str, ids: list[str], documents: list[str], metadatas):
         collection = self.client.get_collection(name)
         collection.add(ids=ids, documents=documents, metadatas=metadatas)
+
+    def get_documents(
+        self, collection_name: str, where_filter: Optional[Dict[str, Any]] = None, limit: Optional[int] = None
+    ):
+        """Gets documents from ChromaDB collection based on filters."""
+        documents = self.client.get_collection(collection_name).get(where=where_filter, limit=limit)
+        return documents
 
     # def query_collection(self, query):
     #     collection = self.client.get_collection(query.collection_name)
